requirements.txt
google-cloud-bigquery
matplotlib
pydantic
pypinfo